+++ angular/lib/src/compiler/identifiers.dart
@@ -143,8 +143,6 @@ class Identifiers {
     new CompileIdentifierMetadata<dynamic>(
         name: "interpolate9", moduleUrl: APP_VIEW_UTILS_MODULE_URL),
   ];
-  static final castByValue = new CompileIdentifierMetadata<dynamic>(
-      name: "castByValue", moduleUrl: APP_VIEW_UTILS_MODULE_URL);
   static final EMPTY_ARRAY = new CompileIdentifierMetadata<dynamic>(
       name: "EMPTY_ARRAY", moduleUrl: APP_VIEW_UTILS_MODULE_URL);
   static final EMPTY_MAP = new CompileIdentifierMetadata<dynamic>(