+++ angular/lib/src/core/linker/app_view_utils.dart
@@ -338,21 +338,14 @@ bool mapLooseIdentical<V>(Map<String, V> m1, Map<String, V> m2) {
   return true;
 }
 
-T castByValue<T>(dynamic input, T value) => input as T;
-
-// TODO(matanl): Remove this hack.
-//
-// Purposefully have unused (_, __) parameters in the proxy functions below in
-// order to get around the lack of a properly typed API for DDC. Once we have a
-// typed API we can remove pureProxy and these unused parameters.
-
 const EMPTY_ARRAY = const [];
 const EMPTY_MAP = const {};
+
 T Function(S0) pureProxy1<T, S0>(T Function(S0) fn) {
   T result;
   var first = true;
   S0 v0;
-  return ([S0 p0, _, __]) {
+  return (S0 p0) {
     if (first || !looseIdentical(v0, p0)) {
       first = false;
       v0 = p0;
@@ -367,7 +360,7 @@ T Function(S0, S1) pureProxy2<T, S0, S1>(T Function(S0, S1) fn) {
   var first = true;
   S0 v0;
   S1 v1;
-  return ([S0 p0, S1 p1, _, __]) {
+  return (S0 p0, S1 p1) {
     if (first || !looseIdentical(v0, p0) || !looseIdentical(v1, p1)) {
       first = false;
       v0 = p0;
@@ -384,7 +377,7 @@ T Function(S0, S1, S2) pureProxy3<T, S0, S1, S2>(T Function(S0, S1, S2) fn) {
   S0 v0;
   S1 v1;
   S2 v2;
-  return ([S0 p0, S1 p1, S2 p2, _, __]) {
+  return (S0 p0, S1 p1, S2 p2) {
     if (first ||
         !looseIdentical(v0, p0) ||
         !looseIdentical(v1, p1) ||
@@ -407,7 +400,7 @@ T Function(S0, S1, S2, S3) pureProxy4<T, S0, S1, S2, S3>(
   S1 v1;
   S2 v2;
   S3 v3;
-  return ([S0 p0, S1 p1, S2 p2, S3 p3, _, __]) {
+  return (S0 p0, S1 p1, S2 p2, S3 p3) {
     if (first ||
         !looseIdentical(v0, p0) ||
         !looseIdentical(v1, p1) ||
@@ -433,7 +426,7 @@ T Function(S0, S1, S2, S3, S4) pureProxy5<T, S0, S1, S2, S3, S4>(
   S2 v2;
   S3 v3;
   S4 v4;
-  return ([S0 p0, S1 p1, S2 p2, S3 p3, S4 p4, _, __]) {
+  return (S0 p0, S1 p1, S2 p2, S3 p3, S4 p4) {
     if (first ||
         !looseIdentical(v0, p0) ||
         !looseIdentical(v1, p1) ||
@@ -462,7 +455,7 @@ T Function(S0, S1, S2, S3, S4, S5) pureProxy6<T, S0, S1, S2, S3, S4, S5>(
   S3 v3;
   S4 v4;
   S5 v5;
-  return ([S0 p0, S1 p1, S2 p2, S3 p3, S4 p4, S5 p5, _, __]) {
+  return (S0 p0, S1 p1, S2 p2, S3 p3, S4 p4, S5 p5) {
     if (first ||
         !looseIdentical(v0, p0) ||
         !looseIdentical(v1, p1) ||
@@ -495,7 +488,7 @@ T Function(S0, S1, S2, S3, S4, S5, S6)
   S4 v4;
   S5 v5;
   S6 v6;
-  return ([S0 p0, S1 p1, S2 p2, S3 p3, S4 p4, S5 p5, S6 p6, _, __]) {
+  return (S0 p0, S1 p1, S2 p2, S3 p3, S4 p4, S5 p5, S6 p6) {
     if (first ||
         !looseIdentical(v0, p0) ||
         !looseIdentical(v1, p1) ||
@@ -531,7 +524,7 @@ T Function(S0, S1, S2, S3, S4, S5, S6, S7)
   S5 v5;
   S6 v6;
   S7 v7;
-  return ([S0 p0, S1 p1, S2 p2, S3 p3, S4 p4, S5 p5, S6 p6, S7 p7, _, __]) {
+  return (S0 p0, S1 p1, S2 p2, S3 p3, S4 p4, S5 p5, S6 p6, S7 p7) {
     if (first ||
         !looseIdentical(v0, p0) ||
         !looseIdentical(v1, p1) ||
@@ -570,8 +563,7 @@ T Function(S0, S1, S2, S3, S4, S5, S6, S7, S8)
   S6 v6;
   S7 v7;
   S8 v8;
-  return (
-      [S0 p0, S1 p1, S2 p2, S3 p3, S4 p4, S5 p5, S6 p6, S7 p7, S8 p8, _, __]) {
+  return (S0 p0, S1 p1, S2 p2, S3 p3, S4 p4, S5 p5, S6 p6, S7 p7, S8 p8) {
     if (first ||
         !looseIdentical(v0, p0) ||
         !looseIdentical(v1, p1) ||
@@ -614,18 +606,17 @@ T Function(S0, S1, S2, S3, S4, S5, S6, S7, S8, S9)
   S8 v8;
   S9 v9;
   return (
-      [S0 p0,
-      S1 p1,
-      S2 p2,
-      S3 p3,
-      S4 p4,
-      S5 p5,
-      S6 p6,
-      S7 p7,
-      S8 p8,
-      S9 p9,
-      _,
-      __]) {
+    S0 p0,
+    S1 p1,
+    S2 p2,
+    S3 p3,
+    S4 p4,
+    S5 p5,
+    S6 p6,
+    S7 p7,
+    S8 p8,
+    S9 p9,
+  ) {
     if (first ||
         !looseIdentical(v0, p0) ||
         !looseIdentical(v1, p1) ||